+++ metrics/RELEASES.md
@@ -37,3 +37,13 @@ Based on:
 - [typescript v0.2.0] metrics
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/@avalanche-sdk/metrics/v/0.2.0 - metrics
+
+## 2025-03-25 18:37:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.523.2 (2.560.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.0] metrics
+### Releases
+- [NPM v0.3.0] https://www.npmjs.com/package/@avalanche-sdk/metrics/v/0.3.0 - metrics